==> notebooks/5. Get and cluster podcasts descriptions embeddings.ipynb <==
# Get and cluster podcasts descriptions embeddings

## Prepare environment

!pip install -q pickle5 podcast-api html2text sentence-transformers

from google.colab import drive
drive.mount("/content/drive")

%cd /content/drive/MyDrive/Coursework

from sentence_transformers import SentenceTransformer
import numpy as np

emb_model = SentenceTransformer('paraphrase-xlm-r-multilingual-v1')

import re
import os
import pickle5 as pickle
import pandas as pd

from copy import copy
from tqdm.notebook import tqdm
from listennotes import podcast_api
from credentials import listennotes_key

client = podcast_api.Client(api_key=listennotes_key)

import re
import html2text

h = html2text.HTML2Text()

h.unicode_snob = True
h.use_automatic_links = True
h.ignore_links = False
h.emphasis_mark = ""
h.strong_mark = ""

def dehtmlify(qhtml, h=h, despacify=True):
    res = h.handle(qhtml)
    res = re.sub(r"\[.*?\(javascript:void\\\(0\\\)\).*?\]", " ", res)
    res = re.sub(r":(\W)", r": \1", res)
    res = re.sub(r"!\[\]\((.*?)\)", r"\1", res)
    res = re.sub(r"(https?:)\s", r"\1", res)
    res = res.replace("\.", ".")
    if despacify:
      res = re.sub(r"\s+", " ", res).strip()
    return res

## Get descriptions

with open("jingles_df.pkl", "rb") as inp:
  jingles_df = pickle.load(inp)
jselect_df = jingles_df.loc[jingles_df["j_status"].isin(["+", "2", "!!", 2])].reset_index(drop=True)
jdf = pd.read_csv("collection_final.csv")
jdf = jselect_df.loc[[list(jselect_df["title"]).index(t) for t in list(jdf["title"])]].reset_index(drop=True)
jdf

#!mkdir jingles

dstruct = {}

for pid in tqdm(list(jdf["ln_id"])):

  response = client.fetch_podcast_by_id(
    id = pid,
    sort = 'recent_first'
  )
  rjson = response.json()

  descriptions = list(set([dehtmlify(rjson["description"])] + [dehtmlify(e["description"]) for e in rjson["episodes"]]))
  dstruct[pid] = copy(descriptions)

with open("dstruct.pkl", "wb") as outp:
  pickle.dump(dstruct, outp)

## Get embedding centres

embs = [np.mean(emb_model.encode(dstruct[pid]), axis=0) for pid in tqdm(list(jdf["ln_id"]))]

## Get clusters

from sklearn.cluster import KMeans

clusterer = KMeans(n_clusters=8, random_state=42)
clusterer.fit(embs)

labels = clusterer.labels_
labels

df = pd.read_csv("collection_final.csv")
df["embeddings_clusters"] = labels
df.to_csv("collection_final.csv", index=False)
