==> notebooks/7. Analyze the results in R.ipynb <==
# Analyze the results in R

install.packages(c("clusteval", "fpc", "stargazer"))

library(clusteval)
library(nnet)
library(stargazer)

jdf <- read.csv("collection_final.csv")
jdf$cat <- as.numeric(factor(jdf$category))

## Cluster similarity test

Between Yandex categories and Sentence Transformers embedding clusters

cluster_similarity(jdf$cat, jdf$embeddings_clusters)

Between Yandex categories and jingles' clusters

cluster_similarity(jdf$cat, jdf$kmeans_clusters)

Between Sentence Transformers embedding clusters and jingles' clusters

cluster_similarity(jdf$embeddings_clusters, jdf$kmeans_clusters)

## Multinomial regression on important measures

mlm <- multinom(category ~ harmony_mean + harmony_var + tempo + mfcc11_mean + mfcc11_var + mfcc15_var + mfcc16_var + mfcc20_mean + mfcc20_var, data = jdf)

summary(mlm)

stargazer(mlm, type="text")
